+++ puzzle/src/main/java/com/example/puzzle/AnimationGamePuz.java
@@ -107,7 +107,7 @@ public class AnimationGamePuz {
 		sprite = spriteJump;
 	}
 
-	public void drawingAnimation(GraphicsPuz graphicsPuz, int x, double y) {
+	public void drawingAnimation(GraphicsPuz graphicsPuz, double x, double y) {
 		graphicsPuz.drawTexture(sprite, x, y);
 	}
